个人中心 管理员

+++ src/main/java/com/example/mallproject/service/RoleService.java
@@ -2,9 +2,6 @@ package com.example.mallproject.service;
 
 import com.example.mallproject.entity.Role;
 import com.baomidou.mybatisplus.extension.service.IService;
-import com.example.mallproject.entity.User;
-
-import java.util.List;
 
 /**
  * <p>
@@ -15,6 +12,6 @@ import java.util.List;
  * @since 2023-04-12
  */
 public interface RoleService extends IService<Role> {
-    Role getUsersAndRoleById(int id);
-    Role getPermissionAndRoleById(int id);
+    Role getUsersAndRoleById(long id);
+    Role getPermissionAndRoleById(long id);
 }
